modify sample-02

+++ sample-02/src/main/java/io/github/devasterisk/springboot/api/repository/UserRepository.java
@@ -8,4 +8,4 @@ import io.github.devasterisk.springboot.api.entity.User;
  * @author Justin Park (aka.asterisk@gmail.com)
  * @since 2018-10-01
  */
-public interface UserRepository extends CrudRepository<User, Long> {}
+public interface UserRepository extends CrudRepository<User, Integer> {}
